+++ tests/test_mdup.cpp
@@ -22,8 +22,8 @@ Test(test_mdup, test_dup)
 
     avm::mdup(&stack);
 
-    cr_assert_eq(4, std::stod(stack.at(0)->toString()));
+    cr_assert_eq(4, std::stold(stack.at(0)->toString()));
     cr_assert_eq(avm::eOperandType::BIGDECIMAL, stack.at(0)->getType());
-    cr_assert_eq(4, std::stod(stack.at(1)->toString()));
+    cr_assert_eq(4, std::stold(stack.at(1)->toString()));
     cr_assert_eq(avm::eOperandType::BIGDECIMAL, stack.at(1)->getType());
 }
